rpcc: Allow setting write buf size and using compression
The Chrome Debug Protocol does not seem to handle websocket
fragmentation properly, for larger messages we have to increase the
write buffer size to prevent Chrome from suddenly terminating the
connection.

There also seems to be an upper limit after which Chrome also terminates
the connection, this happens at 1MB (1048576 bytes), if we subtract the
websocket headers (2 + 8 + 4) this leaves us with 1048562 bytes.

It might be preferable to return an error to the user when the message
size exceeds this limit, however, if compression is enabled we have no
way of knowing at what size this limit is triggered.

Fixes #4.