rpcc: Remove hard-coded 1MB write buffer size limit
The 1MB message size limit was recently bumped to 100MB in Chrome, this
means there is no reason to enforce the 1MB limit any longer.

See: https://github.com/ChromeDevTools/devtools-protocol/issues/24 and
https://chromium-review.googlesource.com/c/560777/

This limitation was in place to improve the user experience (guard
against abrupt disconnection). However, the user is still guarded by the
default buffer size of 4096, writes larger than this will return an
error asking the user to raise the buffer size. These limitations have
now been documented in WithWriteBufferSize so that they can easily be
discovered.