Add WebSocketDebuggerURL to devtool.Version struct
In Chrome >= 62, the "/devtools/browser" websocket endpoint was
removed and replaced with a "/devtools/browser/<UUID>" endpoint [1].
The UUID is generated when the process starts but can be obtained
from the "/json/version" URL:

    $ curl -s https://localhost:9222/json/version
    {
       "Browser": "HeadlessChrome/63.0.3236.7",
       "Protocol-Version": "1.2",
       "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/63.0.3236.7 Safari/537.36",
       "V8-Version": "6.3.272",
       "WebKit-Version": "537.36 (@a1d99e4bc8bf4b58b0717cb406cfe70a9977e9ad)",
       "webSocketDebuggerUrl": "ws://localhost:9222/devtools/browser/74ffefaa-3f6b-4d25-8fb3-98b85d7bf1cd"
    }

This patch just adds the WebSocketDebuggerURL struct member to
devtool.Version so I can get at it from calling code.

There's also some logic in devtool.headlessCreateURL that uses the
"/devtools/browser". I could take a crack at adapting it but wanted
to see if I'm moving in a direction you'd be interested in.

[1]: <https://groups.google.com/d/msg/chrome-debugging-protocol/qElgNKH6ypc/96HDbVV2BwAJ>