rpcc: Fix stream sync deadlock (#91)
* rpcc: Reproduce Sync/notify deadlock from #90

* rpcc: Fix Sync/notify deadlock

* rpcc: Disallow passing same stream twice in Sync

This prevents another deadlock where if the same Stream is passed
multiple times to Sync, Sync would try to acquire the mutex lock a
second time while the first one was still active.

* rpcc: Handle edge case during close in syncMessageStore

Becase (*Conn).notify now calls stream.write outside the streamMu lock,
it's now possible for a write to happen after the syncMessageStore has
closed, in which case there would be no handler for the write and it
could panic. At least theoretically, might need to test this assumption
further.

* rpcc: Ensure continued delivery when closing synced Streams

Fixes #90.