Cleanup session.Manager's handling of errors.
It turns out that cdp.ErrorCause tracks the list of causes and
returns the true cause of the problem. In the case of bug #120,
the topmost err was internal.OpError, and the ErrorCause returns
underlying websocket.CloseError.

Actually there's rpcc.closeError in the middle of cause chain,
but that's simply skipped, and therefore it fails to recognize
the error as closing.

This tracks of every step of Cause and return true if any of
these errors are closing one.