rpcc: Fix potential deadlock in stream Sync
The simplified Conn mutex handling introduced in #74 incorrectly moved
the mutex Unlock to after stream.write. This created the opportunity for
a racy mutex deadlock scenario with simultaneous writes to stream while
`Sync` is being performed.

Due to it's racyness, this was only occasionally caught by
`TestStreamSyncNotifyDeadlock` and evaded detection.